rename js path folders to dash-case instead of wrong-camel
