+++ js/analytics-state-machine/HTML5AnalyticsStateMachine.js
